Add optimistic factory for singletons and provides singleton. Clean up factory tests.
