+++ toothpick-compiler/src/test/java/toothpick/compiler/factory/OptimisticFactoryForSingletonsTest.java
@@ -1,0 +1,30 @@
+package toothpick.compiler.factory;
+
+import com.google.common.base.Joiner;
+import com.google.testing.compile.JavaFileObjects;
+import javax.tools.JavaFileObject;
+import javax.tools.StandardLocation;
+import org.junit.Test;
+
+import static com.google.common.truth.Truth.assert_;
+import static com.google.testing.compile.JavaSourceSubjectFactory.javaSource;
+
+public class OptimisticFactoryForSingletonsTest extends BaseFactoryTest {
+  @Test
+  public void testOptimisticFactoryCreationForSingleton() {
+    JavaFileObject source = JavaFileObjects.forSourceString("test.TestOptimisticFactoryCreationForInjectedField", Joiner.on('\n').join(//
+        "package test;", //
+        "import javax.inject.Inject;", //
+        "import javax.inject.Singleton;", //
+        "@Singleton", //
+        "public class TestOptimisticFactoryCreationForInjectedField {", //
+        "}"));
+
+    assert_().about(javaSource())
+        .that(source)
+        .processedWith(ProcessorTestUtilities.factoryAndMemberInjectorProcessors())
+        .compilesWithoutError()
+        .and()
+        .generatesFileNamed(StandardLocation.locationFor("CLASS_OUTPUT"), "test", "TestOptimisticFactoryCreationForInjectedField$$Factory.class");
+  }
+}
